+++ src/main/java/de/maxhenkel/camera/CreativeTabEvents.java
@@ -2,8 +2,8 @@ package de.maxhenkel.camera;
 
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 
 public class CreativeTabEvents {