+++ src/main/java/de/maxhenkel/camera/ImageCloningRecipe.java
@@ -11,7 +11,7 @@ import net.minecraft.world.inventory.CraftingContainer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.common.crafting.IShapedRecipe;
+import net.neoforged.neoforge.common.crafting.IShapedRecipe;
 import org.jetbrains.annotations.Nullable;
 
 public class ImageCloningRecipe implements CraftingRecipe, IShapedRecipe<CraftingContainer> {