+++ src/main/java/de/maxhenkel/camera/net/MessageDisableCameraMode.java
@@ -5,8 +5,8 @@ import de.maxhenkel.corelib.net.Message;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.network.CustomPayloadEvent;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.neoforge.network.NetworkEvent;
 
 public class MessageDisableCameraMode implements Message<MessageDisableCameraMode> {
 
@@ -20,7 +20,7 @@ public class MessageDisableCameraMode implements Message<MessageDisableCameraMod
     }
 
     @Override
-    public void executeServerSide(CustomPayloadEvent.Context context) {
+    public void executeServerSide(NetworkEvent.Context context) {
         for (InteractionHand hand : InteractionHand.values()) {
             ItemStack stack = context.getSender().getItemInHand(hand);
             if (stack.getItem().equals(Main.CAMERA.get())) {