+++ src/main/java/de/maxhenkel/camera/net/MessageImageUnavailable.java
@@ -3,8 +3,8 @@ package de.maxhenkel.camera.net;
 import de.maxhenkel.camera.TextureCache;
 import de.maxhenkel.corelib.net.Message;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.network.CustomPayloadEvent;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.neoforge.network.NetworkEvent;
 
 import java.awt.image.BufferedImage;
 import java.util.UUID;
@@ -27,7 +27,7 @@ public class MessageImageUnavailable implements Message<MessageImageUnavailable>
     }
 
     @Override
-    public void executeClientSide(CustomPayloadEvent.Context context) {
+    public void executeClientSide(NetworkEvent.Context context) {
         TextureCache.instance().addImage(imgUUID, new BufferedImage(1, 1, BufferedImage.TYPE_INT_RGB));
     }