+++ src/main/java/de/maxhenkel/camera/net/MessageRequestImage.java
@@ -4,8 +4,8 @@ import de.maxhenkel.camera.ImageTools;
 import de.maxhenkel.camera.Main;
 import de.maxhenkel.corelib.net.Message;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.network.CustomPayloadEvent;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.neoforge.network.NetworkEvent;
 
 import java.io.IOException;
 import java.util.UUID;
@@ -28,7 +28,7 @@ public class MessageRequestImage implements Message<MessageRequestImage> {
     }
 
     @Override
-    public void executeServerSide(CustomPayloadEvent.Context context) {
+    public void executeServerSide(NetworkEvent.Context context) {
         try {
             byte[] data = ImageTools.toBytes(Main.PACKET_MANAGER.getExistingImage(context.getSender(), imgUUID));
             Main.SIMPLE_CHANNEL.reply(new MessageImage(imgUUID, data), context);