+++ src/main/java/de/maxhenkel/camera/net/MessageRequestUploadCustomImage.java
@@ -6,8 +6,8 @@ import de.maxhenkel.corelib.net.Message;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.network.CustomPayloadEvent;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.neoforge.network.NetworkEvent;
 
 import java.util.UUID;
 
@@ -29,7 +29,7 @@ public class MessageRequestUploadCustomImage implements Message<MessageRequestUp
     }
 
     @Override
-    public void executeServerSide(CustomPayloadEvent.Context context) {
+    public void executeServerSide(NetworkEvent.Context context) {
         ServerPlayer player = context.getSender();
         if (Main.PACKET_MANAGER.canTakeImage(player.getUUID())) {
             if (CameraItem.consumePaper(player)) {